Add files via upload
This pull request resolves issue #1064 by adding a feature that allows users to convert a binary number to a decimal number in Ruby. Users can input a binary number, and the program validates it before performing the conversion. The decimal equivalent is then displayed to the user. This enhancement improves the usability of the Ruby program.

Signed-off-by: Abhishek Sahay <128229653+abhi5hek001@users.noreply.github.com>